Major tabular syntax change
Any operation that creates scope (i.e., has a rectangular "enclosure" to its
right and below) now requires that the cell ends in ":".  This helps
simplify parsing, and also makes possible the ability to have "bare
tables": tables that don't have any operator (e.g. "table:") and aren't
assigned to any symbol (e.g. "VERB:").

That means that many ordinary plaintext CSV databases are now valid
Gramble source, which makes it much easier to "port" existing projects to
Gramble, and easier for novices to start.  We don't necessarily have to
teach things like ":" ops and assignment in lesson 1, those things can be
introduced when they become relevant.

Also in this update is a bugfix for WebKit/Javascript core (the iOS
Javascript interpreter).  It turns out that WebKit hoists references in a
non-standard way, and so the redeclaration of "var tape" in
State.dQuery() was actually hiding the function argument "tape", and
causing it to be undefined when first used.  Renaming the second
declaration fixes this.