Removed all table obliteration warnings/tests
In the last commit, I removed table obliteration warnings for bare tables,
but not for tables defined with the "table:" op.  But declaring a "table:"
right after other content now, like a bare table, has potential legitimate
uses.

Detecting "lost" code may be a potential kind of check we can run in the
future, but it won't be so simple as saying "Tables can't have prior
siblings", it will require more sophisticated analysis of the source now.
`