Bugfix of nested drop/rename
Encountered a bug when you tried to enclose a DropState or ProjectionState
inside a RenameState.  (This bug is related to the earlier bug that
required Rename "unwrapping", it's just that it turns out that solution
was incomplete.)

I had thought that, when RenameState wraps up its query tape in a
RenamedTape, it could be assured that what it got back was also a
RenamedTape.  (After all, RenamedTapes take care of matchTape to make sure
that the matched tape is also a RenamedTape.)  However, there's an
exception: DropState and ProjectState instead can return new, trivial
TapeCollections.  When RenamedState was trying to unwrap these, it
resulted in an undefined tape, leading to an error in dQuery.  Now that's
fixed.

Meanwhile, slight change for API clarity.  I noticed that the only time we ever called numTapes was to see
if it was empty (meaning, it's a trivial collection).  That makes the code
a little harder to understand ("Why do I have to count the tapes here?"),
so now that getter is replaced by isTrivial, which is more in line with
how we use it.