Factor out Project from sheet parsing
Prior to this commit, we had two very different pieces of functionality
in sheetParser.ts:

  (1) The functionality to parse sheets into States
  (2) A class Project that keeps track of (potentially multiple) sheets
  and their associated namespaces, and also acts as a Facade (in the GoF
  sense) to the various subsystems (so that clients don't have to
  understand where different aspects of the system reside and how they
  interact.)

The only thing preventing their conceptual separation into separate
modules was that the big sheet parsing function, parseCells(), was a
method of Project.  So this commit factors that out to an ordinary function, letting us separate
sheet parsing and Project into two different modules.