Another Namespace bugfix
In writing tests for the previous commit, I found another bug.  It was possible, in attempting to resolve a reference to a symbol in another
sheet, to end up with a reference to a symbol on your own sheet!

How this happens:

  (1) We have loaded Sheet1 and Sheet2.  Assume Sheet1 is the default
  namespace (i.e., the last one explicitly loaded).  Sheet1 defines
  "VERB", and Sheet2 does not.

  (2) We ask for Sheet2.VERB.  The global namespace recognizes this as a
  reference in the Sheet2 namespace, and sends off a request to it to
  resolve VERB.

  (3) This fails, so the namespace starts looking around for other places
  it might find it.  As a last resort, it asks its parent, the global
  namespace.

  (4) The global namespace can't find VERB (it has no symbols at all) but
  it DOES have a default namespace, Sheet1.  It asks Sheet1 to resolve
  VERB, Sheet1 finds it, and returns it.

The error here is that we shouldn't allow the child to call
parent.resolveName() when it was the parent itself who called
child.resolveName().  So now resolveName() takes a param to say whether or
not it's a valid move to ask one's parent as a last resort.