GSuite interface improvements
Three quality-of-life improvements:

(1) You can now sample/generate to a new sheet from the sidebar.  To avoid
the timeout on GAS scripts (that we used to run into when generating),
this generates on the client side (in the sidebar) and then pushes the
results to the GAS backend.

(2) Successful unit tests now highlight in green, so that the programmer
doesn't have to wonder "Did this test succeed, or not execute at all?"

(3) The comment/uncomment functionality in the Sheets add-on menu is
restored.