Porting MatchState back from Python
This commit ports Darlene's implementation of MatchState from the Python
port, eventually intended to be a building block of phonological
replacement rules and similar constructions.

I have not yet ported the whole related test suite; I ported up to the
first failing test and am tracking that down first.

I also implemented some new tests in other suites to try to track down the
issue.  None of these avenues ended up being the issue, but they're useful
tests of some edge cases nonetheless.

One bug that they revealed, and I fixed, is:

In the bizarre case that the programmer has written a grammar with no
literals at all -- just things like Empty(), Any(), etc. -- there are only
two possibilities, that the grammar generates [] or that it generates
[{}].  (There aren't any characters with which to generate any further
output!)  However, we could not always generate {} because some such
grammars would cause an exception.  It came down to a violation of the
assumption that we never query using a tape variable where
tape.isTrivial==true.  This situation violates that because, when there is
no vocabulary on any tape, allTapes.isTrivial==true in State.generate(),
yet we were still trying to send it to ndQuery().  So now we quickly check
that this isn't the case before attempting generation.