New implementation of ConcatState
A bug in nested concatenations of different tapes motivated a big change
to how ConcatStates are implemented.  Basically, the issue came down to
the question "For a concat state A+B, where tape T is relevant to A but
not to B, if A is accepting, is T relevant to B?"  There are some
situations where it has to be (or else you can add additional Ts to B that
A didn't and won't match, leading to incorrect outputs of joins), but
there are some situations where it shouldn't be.  There's possibly a way
to distinguish these situations to decide, but ConcatState is honestly
pretty weirdly complex anyway and my feeling is it's better to replace it
with an alternative formulation now rather than let it get weirder.

The alternative formulation is conceptually much simpler.  The children of
the Concat are stored in an array, we keep an index into this array for
each tape, and we update both indices and the array as needed.

Since ConcatState touches almost everything, changing it caused a (or in
some cases simply revealed) some bugs, not all of which are fixed yet.

The most notable other change is that the new ConcatState is incompatible
with the implementation of Drops and Projections.  (If there's activity
inside a child that the ConcatState doesn't know about, it can't keep its
indexes correct... but that's exactly what Drop/Project do.)  But for some
time we had been thinking to change their implementation, anyway, because
we knew it would eventually cause bugs with how vocabularies are
implemented.  (Even if you hide them with Drop/Project, those tapes are
still "visible" to the vocabulary system under their original names; they have to be or they'd have
no vocabulary.  But then two tapes with the same name, even if dropped,
share a vocabulary.  That's probably not what programmers expect, since
Drop/Project feels like encapsulation and this isn't encapsulation.  So
this would inevitably cause errors that it'd be hard to explain to
programmers, especially in multi-programmer projects.)  So instead, we're
now implementing Drops as name-mangling a la Python private member
variables, using Rename instead.  That way the dropped tape *is* visible
outside the drop to systems that need to see it (like ConcatState and the
vocab system) but will never share a name with any other tape or dropped
tape elsewhere in the grammar.