Change to rename semantics
Previously, Rename did not do anything in particular when it received a named tape query regarding its *from* tape.  E.g., if a RenameState renames "t1" to "t2", but then receives a query regarding "t1", it didn't do anything special, it just forwarded that on to its child.  (I didn't have a strong intuition about what it *should* do in that case.)

However, changing Drop/Project to be name-mangling via Rename leads to reasonable situations that use such a configuration, and there's an obvious answer what the result should be, but we weren't getting that result.  Consider using "Drop" to encapsulate a module such that, inside the module, there's a field called "unit", but it's not visible outside.  Now consider joining that with another grammar where "unit" is defined.  Minimally, something like

  Join(unit("y"), Drop(unit("x"), "unit"))

This should obviously result in {"unit":"y"}; from the outside P.O.V. there's not even a field "unit" on the right hand side.  So forwarding the join's query on "unit" to the child of Drop isn't the right behavior.

In this commit, we define the result of a query on tape T1 to a state Rename(T1 to T2, child) to be a trivial success -- that is, yields [tape, target, false, this].