AnyCharState.accepting bugfix
A few commits ago, accepting() became parameterized by a tape, to be able to ask whether a state was accepting with respect to *that* tape.  I mistakenly thought that AnyCharState.accepting() remained the same -- that like the old AnyCharState, it was always false.  But actually it needs to be the same as LiteralState and be true when the state's tape and the query's tape are different.