Integrate new projection semantics with sheet parser
Made several changes related to projection semantics:

(1) The function Proj() was still using the old ProjectionState, which is incompatible with the new ConcatState.  Rewrote it to use Rename, and also renamed Proj/Drop to Reveal/Hide, which I think are more intuitive regarding what the system is really doing with these tapes.

(2) The sheet parser was expecting the old semantics of Drop, so I changed that to reflect the new Rename-based semantics.

(3) I realized we hadn't been testing the "drop" command in testSheetParser, so I added two CSV projects that use that command.