Hide and reveal header commands
Changed the 'drop' header command to 'hide', and made a corresponding 'reveal' command.

Also we now allow slashed values in the cell (e.g. not just "hide:subj", but "hide:subj/obj"), which is handy when you have a lot of them.

Also, we no longer check whether a hide/reveal is referencing tapes that actually exist in the grammar to the left (since the way we were checking that before no longer applies).  We'll probably eventually want to add that back.