Ops, symbols, and sheets are now case-insensitive
Ops (e.g. "table"), symbol assignments and references, and sheet references (e.g. Verbs.FOO) are now case-insensitive.

TODO: Test this on an OS where files are case-sensitive.  I think we'll run into a bug in the text file interface, because it tries to load a CSV file based on the name of the sheet (i.e. looks for Verbs.csv if the reference is Verbs.FOO), and that won't load if the sheet is, say, Verbs.csv but the reference is verbs.FOO.