Property-based testing
Initial implementation of "property-based" or "metamorphic" testing, in
which we randomly generate thousands of grammars and then test properties
that we believe should be true of them.  For example, some operations
should be commutative, others associative, others idempotent, some should always lead to
identity or equivalence with the empty grammar, etc.

This is exposing many bugs, most of them small, but also important test
cases we hadn't thought to test, like nested repetitions with variable
numbers of repetitions.

Accompanying this is a small bugfix to RepetitionState; the
yieldedAlready flag has (so far as I can tell) ceased to be necessary, and
was preventing the appropriate generation of nested repetitions.