Temporary fix for repetition bug
The grammar Semijoin(t1:h, t2:h*) was recursing infinitely, because t2:h*
was always accepting w.r.t. t1:h, and so would construct successors, but
because indices can go to infinity it was blowing the stack.

For our purposes here, we can solve this with caresAbout, although note
that this isn't a complete & correct solution, because UnionStates can be
such that they care about a tape but also lead to this kind of infinite
acceptance on that same tape.  t1:h|t2:h*, for example, cares about t1 but
also can get into the infinite acceptance loop because of t2.

So this isn't a permanent fix; I'll think more about what the semantics of
these kinds of situations should be.