Minimizing dQuery calls
We were calling dQuery many times when it wasn't absolutely necessary, so
I replaced most such calls with ndQuery.  dQuery is now only used in three
places: in the main parsing loop, when compiling into CompiledStates, and
(of course) in NegationState.ndQuery.

Two small additional changes:

(1) I made a simple StarState as an alternative implementation of the
Kleene star, to verify that the repetition bugs we have been finding
lately are not just errors in our particular implementation, but are
deeper in the semantics.  It appears they are (although the particular
loop that becomes infinite differs!).

(2) I moved property testing into its own folder, so that it's no longer
automatically called by npm test.  It's not the sort of thing we want to
run every time we do continuous integration, for example.