Better fix for repetition bug
Abstractly, the repetition problem comes down to the fact that

   Rep(S).ndQuery(T) == Seq(S, Rep(S)).ndQuery(T)

and when S.accepting(T), then:

   Rep(S).ndQuery(T) yields from Rep(S).ndQuery(T)

So we get into an infinite loop.  But unnecessarily: there aren't
additional results down that path, they're the same results you already
have!

So now RepetitionState makes sure that, when the number of repetitions is
infinite, we don't go down this path.  It's easy for RepetitionState to
catch because it occurs exactly when .child is accepting and .child and
.initialChild are equal.