Randomization is now stateless
The previous implementation of randomization (for sampling randomly from a
grammar) was stateful: randomization was something that happened to a
grammar and changed it, although that could also be undone and the
original grammar restored.

This commit changes randomization to a pure function: S.randomize()
returns a new grammar that's a random subgrammar of S, but doesn't change
S itself.

This makes it a bit easier to reason about; we don't have to worry about
how a random S works compared to S; they're the same.

It's still necessary to pass the "random" parameter to ndQuery, however,
because of the need to choose a new randomization of EmbedState children
and RepetitionState children during the query process.  (Otherwise, something like
(a|b)* would only give you results like "aaaaa" or "bbbbbbbbb" but never
"abbaab".)