Isolated important bug in RepetitionState
It turns out RepetitionState does indeed have the same bug that the old
ConcatState did; it's not able to handle:

  Filter(t1:hh+t2:hh, Rep(t1:h+t2:h), 2, 2))

This is because it maintains a single index (for all tapes) rather than a
dict of indices like the new ConcatState does.  So adding one should fix
this.

Also in this commit, some small changes to the property test framework
(the test for equality was moved from outside the test spec to inside it,
so now tests can do other kinds of tests beyond equality, like subset),
and a few new property tests.