Hidden fields now hidden in GSuite interface
So when there are a bunch of hide commands, it doesn't pollute the
programmer's sidebar interface with "__HIDDEN05_text" inputs and results.

Eventually might be nice to make their display optional, for debugging
purposes.

Also a small fix to display imported symbols with the appropriate
capitalization in the symbol dropdown.