Sheet parser bug tests
Found some tricky bugs in the sheet parser, that will require a
significant refactoring of the architecture (for the better, thankfully).

The problem is with "startswith embed" headers (and endswith and contains
as well).  The thing about these headers (and no others) is that they need
to know what tapes are relevant to their symbols.  (Without that, they
don't know what tapes to accept .* on before/after the symbol.)  That
information doesn't necessarily exist at the time that the "startswith
embed" condition is compiled -- for example, if the symbol has yet to be
defined or itself refers to a not-yet-defined symbol.

The only actually good solution is basically that we need an intermediate representation
between the sheet parser and the actual States -- an abstract syntax tree
(AST) independent of its realization as States, but which has enough
information in it to work out things like tapes.  We need something like
this eventually anyway.

(We actually originally had an AST in the very earliest versions, but when Gramble was simpler it turned
out it wasn't doing any work and I discarded it and only used the grammar
of States themselves.  Now it has some work to do.)