Substantial changes to semantics
This commit makes some substantial changes to the semantics -- enough to
pass all our basic tests (not including some unrelated sheet parser
problems) but not completely followed through.  A lot of refactoring will
probably follow because there are some redundancies now.

Big changes

(1) When you query a state on a tape it doesn't care about, it no longer
does "trivial success" (yielding itself as a successor), it just fails.
Why?  Well, consider for a moment what this meant, it meant that, say,

t1:foo == t1:foo + t2:.*

Well, okay, but consider when you have two of these on different apes

t2:run == t2:run + t1:.*

But this would end up meaning:

t1:foo + t2:run = t1:foo + t2:.* + t2:run + t1:.*

... unless we do something to catch this situation!  And we had -- we had
a special return value "matched" so that matching on one of these implicit
.*s acted differently.  Some states accepted these kinds of special
matches (joins and things that act like joins), others didn't
(concatenations and things that act like concatenations).  But in the end
it all got too complicated; it's better to have a consistent semantics
from the get-go and let the Join family handle it in their own way.

(2) Filters and joins now work completely differently, both because of the
previous change and to make them more efficient.  They now keep track of
their children's tapes; they now need to, to know whether to send a query
to a child, because children no longer trivially succeed when they get an
irrelevant query.  This also lets them prioritize shared tapes so that
we're more likely to fail fast, rather than generate unneeded material
that's later going to be filterd out.

The current implementation is inelegant, I'll probably tighten this up
later.

(3) Now that we know our algorithm is basically Bzhezowski's algorithm,
it's been very helpful to refactor our implementation towards other
implementations of this algorithm, so that we can take advantage of the
insights of other papers about this.  This means going back to a list-like
(as opposed to array-like) implementation of concatenation.  This time,
however, I figured out an implementation that avoids the major bug from
before (where concatenation/repetition of multi-tape concatenations was
broken).

(I wanted to do this in part because this bug STILL plagued repetition,
but implementing the same fix we did for the array-based concatenation
would have been a real headache.  Whereas, if we figured out an equivalent
fix for *list*-based concatenation, it'd be easy to apply to repetition.
And it was.)

The problem that the previous list-like ConcatState had was that it was
moving on from its child1 too soon -- when it was done on *some* tapes but not
*all*.  But we don't want to *not* move on, either -- what happens then is
that it can keep consuming things on that tape even though it's already
accepting, which leads to weird results.  What we needed is some way to
say "that thing's dead so far as tape T is concerned -- accepting, but
never will consume again".

That's very much like the old DropState, so I dusted that off -- now,
instead of moving on to child.ndQuery, it moves on to DropState(chil1, T)
+ child.ndQuery.  That doesn't *quite* work (there's some trouble when it
contains different-tape unions that it's not in the right position to
resolve), but it suggested the
solution.

The real solution lies in accepting().  This "drop" operation was always
right after an accepting() call.  In some
implementations of Brz's algorithm, accepting() actually returns epsilon
or null, not true or false respectively, and is concatenated right into the resulting
state.  In our terms, and not including the drop, it would look like:

   Seq(c1.accepting(t), c2.ndQuery(t))

This implicitly does the work of the "if accepting()": this is equivalent
to c2.ndQuery() if accepting() is epsilon (read: true) and equivalent
to null if accepting() is null (read: false).  Now consider it when we add
that drop:

  Seq(c1.accepting(t), drop(c1, t), c2.ndQuery(t))

The fact that x.accepting(t) and drop(x, t) had the same arguments and
kept co-occuring together suggests they're the same operation.  That would
be an operation that fails whenenever accepting() would, but when it succeeds returns drop(c1, t) -- a state
that will always be accepting on t but will never consume anythign off of
it.

I call this operation "dagger" in the code for want of a better name; that
was just the symbol I was using in my notes.

accepting() itself still exists in the code but I'll probably at some
point make it all dagger().  Except for its use inside generate(),
accepting() was always accompanied by a "moving on" to a different tape
(e.g. in filter), suggesting we should really be using dagger()
everywhere.

(4) As real-world Grambles started to use more filters, sampling was
getting unnacceptably slow.  The problem was that randomization was done
early (by taking a random subgrammar for the current grammar, e.g. taking
one element of a union as if it were the only element), but many of these
grammars wouldn't happen to generate anything when filtered.  Once you
have enough filters, it becomes rather rare to sample successfully, and
all those failed tries were taking up too much time.

Now, rather than go for a "sample the grammar" approach, we do a random
search (instead of our old breadth-first search) through actual outputs.
That, combined with the more efficient filtering described above, should
make it much faster in practice.

(5) Finite repetition is now handled without any special state, just as
sequences and Maybes.  There's now only a special state for the actual
Kleene star, the zero-to-Infinity repetitions.  N-to-infinity is just
handled by the child repeated N times followed by the Kleene star.

Small additional changes:

(a) It was still possible that NegationState return empty BitSets (like
LiteralState used to be able to do).  This can lead to bugs now that we
call dQuery only sparingly, so I fixed this.

(b) A bunch more small tests, to test for things that either came up
during the semantics change, or that came up in property testing.

(c) Now that joins actually work again, I added them to property testing.