Dead code removal, fixed rename bug
(1) Removed the whole "matched" system since this was a legacy of old
code.

(2) RenameState still had "trivial success" semantics, that's what had
broken that one rename test a while back. (I thought maybe the test itself
was mistaken and had commented it out.)

(3) Removed the "random" parameter from accepting() and (n)dQuery()
because that's not how randomization works anymore.

(4) Abandoning the old randomization system means that the array-based
UnionState isn't necessary anymore.  I went back to the list-based
implemnetation because it's closer to other Brz algo implementations.

(5) Went through and removed references to the TrivialState (that is,
epsilon) being "0" or
"empty" because it's very important in more traditional Brz algo
implementations to distinguish epsilon from the empty set.  It's now
consistently called epsilon.