Start of abstract syntax tree code
There are various operations (esp. related  to type inference -- in our case, "tape"
inference) where sheet parsing is too early (not all the information is
present yet) but doing it in states would be too late (in some case we
need that information before we make the states).

So we need an intermediate layer -- something that the sheetParser can
create, where we can do type inference and some kinds of error checking,
and that then constructs the States that will do the actual calculation.

We originally did have this layer, but it went away very early, possibly
before the first commit, because at that time it didn't do anything -- it
was just redundant hierarchy mirroring States.  Now as the language is a
lot more complex, it actually has things to do.

This commit lays out a very basic AST hierarchy, but doesn't hook it up to
anything yet.