Corrected Brz semantics for repetition
The Brzozowski semantics (for a 1-tape language) for the Kleene star has
the delta of A* be epsilon.  I wasn't quite sure what the multi-tape
generalization of this was, so for the time being I also implemented it as
being epsilon.  However, actually d^t(A*) = (d^t(A))*.  Since d^t(A) will
always be epsilon or null for a one-tape language, and epsilon* and null*
are both equal to epsilon, then this rule simplifies to Brzozowski's rule
for one-tape languages.

Fixng this allows us to switch to using delta() in the main generate() loop instead of
accepting().  Elsewhere, there are a few places it remains that I'll get to
soon.