Removing accepting() in favor of delta()
As mentioned in the 2021-06-18 commit, accepting() and what was then
called dagger() (now called delta()) aren't both necessary.
accepting() is redundant, in that it's always true when delta() is
anything but null, and false when delta is null.  I mentioned then that I
would eventually eliminate accepting(); this commit does that.

This commit also eliminates TextState, the shared ancestor of LiteralState
and AnyCharState; they're now just direct descendants of State.  This
state only existed to express the children's ndQuery as a template method,
and that referenced accepting() in a way that was awkward to express in
terms of delta().  I think it's clearer what's going on when both states just
have their own ndQuery.