Breadth-first search through tapes
While, within any given tape, our search through the parse space was
breadth-first, we were actually effectively depth-first when considering
the search through the tapes themselves.  (Or, at least, could run into
some of the pathological behavior of depth-first search.)  It was possible
to keep querying for material on one tape, while on another tape we would
have failed at the first character.

Now that we've got a generation algorithm that can correctly manage to
generate to any specific tape at any time, we can effectively implement
Roland's suggestion for breadth-first search at the tape level.

We don't, however, want to go down every delta path every time, because
that would lead to lots of derivational diamonds.  (E.g., going down
delta(t1) and then delta(t2), vs. delta(t2) then delta(t1).  Those have
the exact same effect -- they get rid of t1 and t2 and output nothing --
but there are two derivational pathways to it.)

So instead we just choose whatever the first tape is, query and delta from
that, and each time "rotate" the array so that a new tape is in front.
That way we try each of the tapes once before coming back to the original.