Smart constructors and Dot
Moving "dot" into the Ast/Brz semantics and testing.

There was some complication regarding the special case of empty
vocabularies, the new version of the trivial vocabulary check at the
beginning of generate() didn't work as programmed.  But the new main loop
of generation can actually handle these without a special check, so long
as the states are simplified appropriately.  (E.g., that epsilon+epsilon
is reduced to epsilon.)

This was a good opportunity to switch to using smart constructors rather
than a simplify method.  (Smart constructors are also used by Owens et al.
2009, Regular-expression derivatives reexamined, for the same purpose.)
Rather than create expressions using new X() constructors and then
simplifying them, we use non-method functions createX() that construct the
simplified form from the outset.  (These have to be non-method functions
because the object created might not be of the expected class, for example
creating a concat with epsilon and epsilon results in something of class
BrzEpsilon, not BrzConcat.)

So long as we consistently use createX()
rather than new X(), we'll always be working with simplified expressions.
That's just easier to reason about than worrying about when and whether to
call simplify().