Integrated embed and hide
"Hide" is new in this version; I didn't implement "reveal" because it's
more tedious and no one's using it anyway.

We already had "Embed" but it had a few
outstanding issues:

(1) Our old dynamic name resolution was case-insensitive, but the new
static name resolution was still case-sensitive.  This commit makes it
case-insensitive.

(2) Found and fixed a bug when nested namespaces shared
names.  When a namespace named X contains a namespace named X, and both
have a symbol named Y, then X.Y would always refer to the outer one,
whereas it should refer to the inner.  We were forgetting to break in
qualifyNames() when we successfully resolved a name, and so it kept
looking even once it had found one.  (Meaning, we were always finding the
*last* possible referent, the outermost one, rather than the *first*
possible referent, the innermost one.)  This commit fixes this.