Multi-tape negation
Ported negation to the new semantics, and while I was at it made it
multi-tape; previously it only allowed negations of single-tape grammars.

Negation poses a bit of a puzzle to the multi-tape semantics.  We have a
rule that, for any grammar X, if tape T isn't defined anywhere in X, then
its value is epsilon.  What about the negation of X?  Is the value of T
~epsilon (aka .+)? I suppose so, but it messes up the algorithm as
written.  (And it's kind of hard to think about -- it has additional
material on at least one of infinitely many tapes not defined in the
grammar?)

So for the purposes of Gramble, let's say that negation is relative to
some fixed number of tapes -- specifically, the tapes defined within the
child.  Otherwise it's just like every other grammar, it's epsilon on
tapes not in that grammar.  This is good enough for now.