Integrating ASTs with sheet parsing
First pass at integrating the new AST/deriv semantics with sheet parsing.
Right now it's just enough to get the tests passing, there's still a lot
of redundancy between the sheet parser's Component hierarchy and the AST,
which eventually should only be in ASTs.

Working through the sheet parser tests revealed two bugs in the
ASTs/derivs:

(1) Forgot a "continue" in the generate loop after we've found an epsilon,
with the result that some trivial results were being generated twice.

(2) The ASTs weren't collecting vocabulary correctly w.r.t.
Embeds/Namespaces.  The Namespace was considering its vocabulary to be the
result of calling collectVocab() on each symbol, but that's incorrect, it
should be the results of collectVocab() on the last-defined symbol
(similar to how calculateTapes and constructExpr work).

The incorrect old way and the correct new way often have the same results,
which is why the existing tests didn't catch it, it's only when you have
embedding and renames together.  The basic rename/embed tests failed to
test for this, but the sheet parser tests did.

I've added relevant tests to the rename and hide test suites (AST
versions) so that we can catch these even without the sheet apparatus.