Integrating equals/startsWith/etc to ASTs
This finally addresses a longstanding bug, one of the major reasons we
moved to an AST in the first place.

One of the reasons for needing an intermediate AST representation in the
first place was that the sheet parser itself couldn't correctly construct
startsWiths, endsWiths, and contains during the sheet parsing stage.  To
make these, it needs to know the tapes relevant to the condition, but
these can't be known until the condition is actually constructed.  This
led to a bug where startsWiths etc. weren't being correctly constructed
when the condition contains an embed that is defined only after the
startsWith is constructed.

But the "state" objects themselves were too late -- you can't wait until
after all states are constructed, because what you're doing is creating
states.  So an intermediate layer is needed.

The AST hierarchy is now our intermediate layer.  The sheet parser just
creates an AstStartsWith object, not knowing exactly which Brzozowski
expressions that will eventually generate.  Then later, the AST hierarchy
works out what tapes are relevant to each node in the tree, and only after
that creates the actual Brzozowski expressions.