Porting Match from main branch
Porting Darlene's MatchState implementation and tests from the main
branch.

The main change is that we need a delta() implementation rather than
accepting(), and delta() is actually fairly tricky to get right.  The
issue is that, if you have Match(A:a+B:epsilon), delta w.r.t.B will be
null -- not because of the contents of B, but because of the contents of
A, which must match B.  This is the first time where delta(B) depends on
some other tape.

So delta(B) has to call delta(A), but at that point in the call it doesn't
know about A -- it knows the tape name "A" but doesn't have the Tape object
corresponding to it.  So we have to find something that knows about A.
The best candiate for this is the B tape object -- they're siblings within
the tape hierarchy -- so I added a way for Tapes to return their siblings
when given their names.