Major refactoring of sheet parsing
Sheet parsing, over time, had gotten fairly convoluted, with responsibilities shared somewhat haphazardly between different hierarchies of components.  (Especially since ASTs came into the picture, since they took some of the responsibility for representing the logical structure of the grammar.)

This commit attempts to make a clearer division of knowledge/responsibility/etc between the different component hierarchies.

The various components in sheetParser are now all consistently named "Tst-" (tabular syntax tree), and represent the logical structure of the program as typed (e.g., the collection of operators, tables, assignments, etc.).  This is a different kind of representation than the logical structure of the *grammar* itself: various kinds of syntactic structures in Gramble result in [AstSequence]s, for example.  For the purposes of the grammar, all of those are identical and should be treated as sequences -- there's no difference between how the language treats a sequence you made by putting cells in a row, by a slash header, or by a "concat:" operation.  But we still need it: it's important for things like error checking to know how the programmer actually expressed that sequence.  (For example, if there's nothing that follows a "concat:" operation, that's an error and we have to notify the programmer.  If there's nothing that follows a simple row of cells, that's not an error -- or if it is, we'd never have any way of knowing.  By the time these are both [AstSequence]s, this is obviously too late to distinguish these cases.)

So there are now three basic levels of representation.  TSTs express the program as written in the tabular syntax, and are responsible for transforming the program as written into an abstract grammar.  ASTs express the grammar, and are the level at which we do things like type inference (in our case "tape inference") and static name resolution.  [Expr]s are the formulas that the parse/generate algorithm is juggling; this needs to be yet another respresentation because the relationship between the grammar unit and the type of formula that expresses it is complex -- not all [AstSequence]s are [ConcatExpr]s, and [ConcatExpr]s are used in many situations that never correspond to any actual sequence in the grammar.

Changes of note:

(1) We previously called compile() on TSTs; this was a void method that changed the TST's .ast member from Epsilon to whatever it should be.  Since too many things are called "compile" already, this is now called "toAST()", and it returns the AST rather than storing it as a member.  It's not 100% stateless currently, as it also does some error handling as a side effect.  (So if you call .toAST() twice, that error is registered twice, and our error system isn't currently smart enough to realize that's one error.)

(2) It used to be that we created potentially many TST objects corresponding to a single cell -- when, for example, the header and/or cell were complex, like "startswith text:~(n|m)".  (So there'd be a component corresponding to the startswith, it would have a child corresponding to the negation, and so on.)  But there actually wasn't a lot they were doing, and now that there are things like [AstStartsWith], they weren't really doing anything anymore.  Now, each cell (aside from ones that nothing's written in) corresponds to roughly one TST.  (Actually many are two, because cells are [TstCell] before they're added to something else.  And things like the first cell of a [TstTable] end up corresponding to three.  But roughly.)

(3) The hierarchy of headers (the results of parsing a potentially complex header like "startswith text") are no longer TSTs.  [Header]s are now their own file that knows nothing of TSTs, and within the TST hierarchy itself there's now a [TstHeader] that *contains* a [Header].  Like before, though, header objects are still ultimately the ones responsible for turning a header/cell pair like "startswith text:~(n|m)" into the appropriately complex [AstComponent].  In order to do this, after all, you need to understand the structure of "startswith text", and its the Header hierarchy that knows that.

Headers still basically do the "compile your child, then merge it with the left neighbor" thing, but I'm no longer doing that as an explicit template algorithm where each header could implement any of {.compile(), .merge(), .compileAndMerge()}.  It just got hard to reason about what was going on, and in the end didn't get rid of all that much duplicate code.  Now only one lineage of headers retains a separate .merge() function, the equals/startswith/etc. headers, because there it really does eliminate a ton of difficult, repeated code.

(4) Join headers (the ones that look like "@gloss") are no longer a thing.  Some of the refactoring broke them, and I decided "good riddance"; we should be using "equals" there instead.

Some of Fineen's grammars use join headers purely as a way to get an alternation (e.g. to let her say "unit:spring|summer") without having to express this in two separate lines.  This is a good pattern (two lines means repeated information, and repeated information leads to bugs), but we're not actually using the join capability here, just the logic parsing.  So there will eventually be a "logic X" parser that will turn on logic but still merge using Seq(); the object is there but it's not hooked up to anything yet.

(5) Error handling is all over the place now; every hierarchy does it in a slightly different way.  This should eventually be unified into something sensible.